+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Next Release
 
+### Added
+- (Sprite Editor) Copy/paste/duplicate for commands/keyframes
+- Nix compatibility
+
 ### Changed
 - Main config and logs moved to system user directory (%AppData% on Windows, etc)
-- Added nix compatibility
 
 ### Fixed
 - (Map Editor) Fixed crash when fusing vertices